refactor: move utility functions for IBAN, account numbers, and date/time to utils
